# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Bulk Purchasing Platform API
  description: |
    Complete API documentation for the Save Together bulk purchasing application.
    
    ## Features
    - User authentication with JWT tokens
    - Product catalog management
    - Vendor management and dashboards
    - Bulk purchasing groups
    - Order management with MOQ validation
    - Escrow-protected transactions
    - Dispute resolution
    - Real-time chat with WebSocket (Socket.IO)
    
    ## Authentication
    Most endpoints require a JWT token obtained from `/auth/login`. Include the token in the Authorization header:
    ```
    Authorization: Bearer {token}
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@savetogether.com
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://api.savetogether.com
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login endpoint. Format is "Bearer {token}"
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [superUser, admin, vendor, member]
          description: User's global role in the system
        avatar:
          type: string
          format: uri
          nullable: true
        vendor_id:
          type: string
          format: uuid
          nullable: true
          description: ID of vendor if role is 'vendor'
        trust_score:
          type: integer
          nullable: true
        is_active:
          type: boolean
        is_online:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, email, name, role, created_at]
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        image:
          type: string
          nullable: true
        bulk_price:
          type: number
          format: decimal
          description: Price per unit for bulk orders
        retail_price:
          type: number
          format: decimal
          description: Regular retail price
        moq:
          type: integer
          description: Minimum Order Quantity
          minimum: 1
        vendor_id:
          type: string
          format: uuid
        category:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, name, bulk_price, retail_price, moq, vendor_id]
    Vendor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        rating:
          type: number
          format: decimal
          minimum: 0
          maximum: 5
        location:
          type: string
          nullable: true
        image:
          type: string
          format: uri
          nullable: true
        verified:
          type: boolean
        product_count:
          type: integer
        created_at:
          type: string
          format: date-time
      required: [id, name, verified]
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
          description: Unique order identifier
        group_id:
          type: string
          format: uuid
          nullable: true
        buyer_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, paid, confirmed, shipped, delivered, cancelled]
        total_amount:
          type: number
          format: decimal
        estimated_delivery:
          type: string
          format: date-time
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              product_id:
                type: string
                format: uuid
              product:
                $ref: '#/components/schemas/Product'
              quantity:
                type: integer
              price:
                type: number
                format: decimal
        created_at:
          type: string
          format: date-time
      required: [id, order_number, status, total_amount]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
      required: [success, error]
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      security:
        - bearerAuth: []
      description: Admin only. Get all users in the system
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'

    post:
      tags: [Users]
      summary: Create user
      security:
        - bearerAuth: []
      description: Admin only. Create new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                name:
                  type: string
                role:
                  type: string
                  enum: [superUser, admin, vendor, member]
                vendor_id:
                  type: string
                  format: uuid
                  nullable: true
              required: [email, password, name, role]
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
...
  /products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by product category
        - name: vendor_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by vendor
        - name: is_active
          in: query
          schema:
            type: boolean
          description: Filter by active status
      responses:
        '200':
          description: Products list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'

    post:
      tags: [Products]
      summary: Create product
      security:
        - bearerAuth: []
      description: Vendor only. Create new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                bulk_price:
                  type: number
                  format: decimal
                retail_price:
                  type: number
                  format: decimal
                moq:
                  type: integer
                  minimum: 1
                vendor_id:
                  type: string
                  format: uuid
                category:
                  type: string
                image:
                  type: string
              required: [name, bulk_price, retail_price, moq, vendor_id, category]
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Product'
...
  /vendors/{id}:
    get:
      tags: [Vendors]
      summary: Get vendor details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vendor details with products
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      vendor:
                        $ref: '#/components/schemas/Vendor'
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
  /vendors/{id}/dashboard:
    get:
      tags: [Vendors]
      summary: Vendor dashboard stats
      security:
        - bearerAuth: []
      description: Vendor only. Get vendor's business statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vendor dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      total_revenue:
                        type: number
                      monthly_revenue:
                        type: number
                      total_orders:
                        type: integer
                      pending_orders:
                        type: integer
                      total_products:
                        type: integer
                      total_customers:
                        type: integer
                      average_rating:
                        type: number
                      total_reviews:
                        type: integer
  /vendors/{id}/orders:
    get:
      tags: [Vendors]
      summary: Vendor's orders
      security:
        - bearerAuth: []
      description: Vendor only. Get orders for vendor's products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vendor orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
  /vendors/{id}/customers:
    get:
      tags: [Vendors]
      summary: Vendor's customers
      security:
        - bearerAuth: []
      description: Vendor only. Get list of vendor's customers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customers list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        user:
                          $ref: '#/components/schemas/User'
                        total_orders:
                          type: integer
                        total_spent:
                          type: number
                        last_order_date:
                          type: string
                          format: date-time
                        trust_score:
                          type: integer

  # ============ ORDERS ============
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'

    post:
      tags: [Orders]
      summary: Create order
      security:
        - bearerAuth: []
      description: Create new order with items. Validates MOQ per product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    properties:
                      product_id:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        minimum: 1
                group_id:
                  type: string
                  format: uuid
                  nullable: true
              required: [items]
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Validation error or MOQ violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...